refactor(zip): all under one file, move patch file
- moved all zip logic under one file
- moved patch to add/observable/zip

BREAKING CHANGE: add/operator/zip-static is now add/observable/zip
