+++ src/Rx.ts
@@ -27,7 +27,7 @@ import './add/observable/never';
 import './add/observable/range';
 import './add/observable/throw';
 import './add/observable/timer';
-import './add/operator/zip-static';
+import './add/observable/zip';
 
 //operators
 import './add/operator/buffer';
