+++ src/add/observable/zip.ts
@@ -1,0 +1,6 @@
+import {Observable} from '../../Observable';
+import {zipStatic} from '../../operator/zip';
+
+Observable.zip = zipStatic;
+
+export var _void: void;
