--- src/add/operator/zip-static.ts
@@ -1,7 +1,0 @@
-
-import {Observable} from '../../Observable';
-import {zip} from '../../operator/zip-static';
-
-Observable.zip = zip;
-
-export var _void: void;
