--- src/operator/zip-static.ts
@@ -1,11 +1,0 @@
-import {Observable} from '../Observable';
-import {ArrayObservable} from '../observable/ArrayObservable';
-import {ZipOperator} from './zip-support';
-
-export function zip<T, R>(...observables: Array<Observable<any> | ((...values: Array<any>) => R)>): Observable<R> {
-  const project = <((...ys: Array<any>) => R)> observables[observables.length - 1];
-  if (typeof project === 'function') {
-    observables.pop();
-  }
-  return new ArrayObservable(observables).lift<T, R>(new ZipOperator<T, R>(project));
-}
